agent/requirements.txt
# requirements.txt
mysql-connector-python==8.2.0
gymnasium==1.2.3
stable-baselines3==2.7.1
numpy==1.24.3  
torch==2.10.0
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.2
pandas==2.1.4